Be a comfortable guest.
A lot of my local tweaks are not muscle memory -- it's frustrating not
to have them on remote systes. But, in many cases, I don't own the
remote account. This gives the best of both worlds.

To use:
    scp -rp guest remote_user@remote_system:hwine_guest
The each time you log in:
    source hwine_guest/use_hal

Note: the files in "guest/" are symlinks to the normal copies of these
files. The 'scp -rp' will copy current contents of symlinks.